pyproject.toml
[project]
name = "cli-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "gradio>=6.9.0",
    "groq>=1.0.0",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
]
